requirements.txt
pandas==2.2.3
openpyxl==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
supabase>=2.0.0
dateparser==1.1.8
playwright==1.38.0
tqdm==4.66.1